minor debugs

+++ pheromone.go
@@ -13,7 +13,7 @@ func addPheromoneAlongPath(pos1, pos2 Pos) {
 	dx := pos2.x - pos1.x
 	dy := pos2.y - pos1.y
 
-	steps := int(math.Max(math.Abs(dx), math.Abs(dy)))
+	steps := (math.Max(math.Abs(dx), math.Abs(dy)))
 
 	if steps == 0 {
 		// If the positions are the same, just add pheromone to the single pixel
@@ -21,10 +21,10 @@ func addPheromoneAlongPath(pos1, pos2 Pos) {
 		return
 	}
 
-	xStep := dx / float64(steps)
-	yStep := dy / float64(steps)
+	xStep := dx / (steps)
+	yStep := dy / (steps)
 
-	for i := 0; i <= steps; i++ {
+	for i := 0; i <= int((steps)); i++ {
 		x := pos1.x + float64(i)*xStep
 		y := pos1.y + float64(i)*yStep
 		addPheromoneToPixel(int(x), int(y))
@@ -33,7 +33,7 @@ func addPheromoneAlongPath(pos1, pos2 Pos) {
 
 func addPheromoneToPixel(x, y int) {
 	if x >= 0 && y >= 0 && x < ImageWidth && y < ImageHeight {
-		Game.PheromoneImage.Set(x, y, color.RGBA{0, 0, 255 * PheromoneDeposit, 255})
+		Game.PheromoneImage.Set(x, y, color.RGBA{0, 0, uint8(255.0 * PheromoneDeposit), 255})
 	}
 }
